+++ assets/components/radio-input.component.jsx
@@ -13,13 +13,19 @@ radioInput.mixins = [
 
 radioInput.propTypes = {
   className: React.PropTypes.string,
-  options: React.PropTypes.array.isRequired
+  value: React.PropTypes.any,
+  radioValue: React.PropTypes.any.isRequired,
+  label: React.PropTypes.node,
+  labelPosition: React.PropTypes.oneOf(['left', 'right'])
 };
 
 radioInput.getDefaultProps = function radioInputGetDefaultProps() {
   return {
     className: null,
-    options: []
+    value: null,
+    radioValue: null,
+    label: null,
+    labelPosition: 'right'
   };
 };
 
@@ -27,7 +33,8 @@ radioInput.getInputPassThroughProps = function() {
   var props = _.clone(this.props);
 
   delete props.className;
-  delete props.options;
+  delete props.label;
+  delete props.labelPosition;
   delete props.value;
 
   return props;
@@ -44,61 +51,53 @@ radioInput.getCssClasses = function radioInputGetCssClasses() {
 };
 
 radioInput.renderInput = function radioInputRenderInput() {
-  var options = _.map(this.props.options, function radioInputRenderInputOptionsMap(option) {
-    var checked = this.props.value === option.value;
-    var icon = null;
-    var iconProperties = {
-      fragment: 'radio-empty',
-      outerClassName: 'm-right'
-    };
+  var checked = this.props.value === this.props.radioValue;
+  var icon = null;
+  var iconProperties = {
+    fragment: 'radio-empty',
+    outerClassName: 'm-right'
+  };
 
-    if (checked === true) {
-      iconProperties.fragment = 'radio-checked';
-    }
+  if (checked === true) {
+    iconProperties.fragment = 'radio-checked';
+  }
 
-    if (option.displayPosition === 'left') {
-      iconProperties.outerClassName = 'm-left';
-      icon = React.createElement(SvgIcon, iconProperties);
-
-      return (
-        <label
-          className="m-toggle"
-          key={option.value}
-        >
-          {option.display}{icon}<input
-            className="form-element__input m-radio m-left"
-            type="radio"
-            checked={checked}
-            value={option.value}
-            {...this.getInputPassThroughProps()}
-          />
-        </label>
-      );
-    }
-
+  if (this.props.labelPosition === 'left') {
+    iconProperties.outerClassName = 'm-left';
     icon = React.createElement(SvgIcon, iconProperties);
 
     return (
       <label
         className="m-toggle"
-        key={option.value}
+        key={this.props.radioValue}
       >
-        {icon}<input
-          className="form-element__input m-radio m-right"
+        {this.props.label}{icon}<input
+          className="form-element__input m-radio m-left"
           type="radio"
-          value={option.value}
           checked={checked}
+          value={this.props.radioValue}
           {...this.getInputPassThroughProps()}
         />
-        {option.display}
       </label>
     );
-  }.bind(this));
+  }
+
+  icon = React.createElement(SvgIcon, iconProperties);
 
   return (
-    <div className="form-element__radio-group">
-      {options}
-    </div>
+    <label
+      className="m-toggle"
+      key={this.props.radioValue}
+    >
+      {icon}<input
+        className="form-element__input m-radio m-right"
+        type="radio"
+        value={this.props.radioValue}
+        checked={checked}
+        {...this.getInputPassThroughProps()}
+      />
+      {this.props.label}
+    </label>
   );
 };